Protect against NRE
TextOption view model string values aren't really supposed to
be null - they should be the empty string if not present.
That said, with this change if the value is null we
turn that into the empty string rather than generating
an NRE - that's much safer in general and fixes AB#1487328
specifically.